pyproject.toml
[project]
name = "airplane-detection"
version = "0.1.0"
description = "Airplane detection using YOLOv8 on satellite imagery"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "ultralytics>=8.0.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "pyyaml>=6.0",
]

[project.scripts]
train = "scripts.train:main"
evaluate = "scripts.evaluate:main"
predict = "scripts.predict:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]